Relocate "generate" CLI code outside of transaction and improve error handling.
The relocation allows for faster transactional operation.
With the loop outside of the transaction, then the transaction can be committed for each pass.
This then allows for reduced memory requirements for large transactions (such as when generate a huge amount of submissions).

The exception handling is improved by catching the exceptions at the CLI level such that exceptions can be reported without crashing the entire server.
Now, if a CLI fails due to some error, then the server does not exit.