Issue 58: Unable to re-arrange the submission workflow in Manage Workflow.
The organization is being changed but the dirty flag is not being set.
Manually set the dirty flag before sending.

The `OrganizationRepo` listeners are either not doing anything or they are not checking for when status is a failure.
The `ApiResponseActions.READ` and `ApiResponseActions.BROADCAST` never do anything because the `resetManageOrganization()` only does anything when the organization is provided.
No organization is being provided.