Issue 58: Continue to call the resetManageOrganization() in a way that does something.
The `OrganizationRepo.listen(ApiResponseActions.READ,...` call was removed for being a no-op.
Even older code suggests that this should in fact be doing something.
Add the function call back, but this time add the selected organization if one exists so that the function is not a no-op.

Whether or not this change fixes or breaks something is unclear.
Be wary of regressions.
Be hopeful of bonus bug fixes.