Improve CLI generation, fixing embargoes to be actual embargoes and several others.
The embargoes are now generated from the actual list of embargoes.
If there are no embargoes then it falls back to a randomly generated string.

Handle other similar cases that are easy to resolve at the same time.
This makes the generated data more accurate and easier to test.
Especially when using them for filtering.