Issue 1931: Pagination not working properly for combined filters list.
This is a regression from the aggressive performance optimizations such as commit b2c538c5c9c7d99a18c9f8c2c47b8cc98c01a260.

The case that is not correctly handled is where there are multiple separate field values being selected and filtered upon.

This reverts the problematic optimization to ensure that the generated queries are correct.
This has a cost of losing a significant performance optimization.
The performance optimization will have to be re-looked at and a more correct approach without any regressions will need to be figured out.