The "None" filters should handle NULL case when not configured.
If the `EmbargoType` and `SubmissionType` `None` settings are not specified in the `application.yml` or via an environment variable then a `NULL` pointer exception no longer is thrown.
When not configured, then the `None` settings are ignored entirely.
When not configured, then the `None` searches do not match `NULL` but do match the string.